[REF] web, project: extract embedded actions from ControlPanel
Before this commit, the ControlPanel component was coupled with
the logic, state management, and UI rendering for "embedded actions".

This commit separates these concerns by extracting the embedded actions
system into its own files.

This refactoring reduces the complexity of `control_panel.js`
without altering the behavior for views inheriting from it.

Task~4277541
Related~odoo/enterprise#120999
