+++ addons/project/static/src/views/project_task_control_panel/project_task_control_panel.xml
@@ -3,18 +3,18 @@
 
     <t t-name="project.ProjectTaskControlPanel" t-inherit="web.ControlPanel">
         <xpath expr="//button[@name='embedded_actions']" position="replace">
-            <t t-if="this.showTaskOptions and (!this.env.isSmall or !this.state.embeddedInfos.embeddedActions?.length)">
+            <t t-if="this.showTaskOptions and (!this.env.isSmall or !this.embeddedPanelState.embeddedInfos.embeddedActions?.length)">
                 <Dropdown>
                     <button class="btn btn-secondary" t-att-title="this.taskOptionsTitle">
                         <i class="fa fa-sliders"/>
                     </button>
                     <t t-set-slot="content">
-                        <DropdownItem onSelected="() => this.onClickShowSubtasks()" class="this.state.showSubtasks ? 'selected' : ''">
+                        <DropdownItem onSelected="() => this.onClickShowSubtasks()" class="this.embeddedPanelState.showSubtasks ? 'selected' : ''">
                             Show Sub-Tasks
                         </DropdownItem>
-                        <DropdownItem t-if="this.state.embeddedInfos.embeddedActions?.length and !this.env.isSmall"
-                            class="this.state.embeddedInfos.showEmbedded ? 'selected' : ''"
-                            onSelected="() => this.onClickShowEmbedded()"
+                        <DropdownItem t-if="this.embeddedPanelState.embeddedInfos.embeddedActions?.length and !this.env.isSmall"
+                            class="this.embeddedPanelState.embeddedInfos.showEmbedded ? 'selected' : ''"
+                            onSelected="() => this.embeddedPanelState.onClickShowEmbedded()"
                         >
                             Top Menu
                         </DropdownItem>
@@ -22,13 +22,13 @@
                 </Dropdown>
             </t>
         </xpath>
-         <xpath expr="//t[@t-if='this.state.embeddedInfos.embeddedActions?.length and this.env.isSmall']" position="replace">
+         <xpath expr="//EmbeddedActionsPanel[@id='embedded-actions-panel-mobile']" position="replace">
             <t t-call="project.MobileCombinedActionsDropdown"/>
         </xpath>
     </t>
 
    <t t-name="project.MobileCombinedActionsDropdown">
-        <t t-if="this.env.isSmall and this.state.embeddedInfos.embeddedActions?.length">
+        <t t-if="this.env.isSmall and this.embeddedPanelState.embeddedInfos.embeddedActions?.length">
             <Dropdown menuClass="'o_combined_mobile_dropdown'">
                 <button class="btn btn-secondary">
                     <i class="fa fa-sliders" />
@@ -36,16 +36,16 @@
                 <t t-set-slot="content">
                     <DropdownItem
                         onSelected="() => this.onClickShowSubtasks()"
-                        class="this.state.showSubtasks ? 'selected' : ''"
+                        class="this.embeddedPanelState.showSubtasks ? 'selected' : ''"
                     >
                         Show Sub-Tasks
                     </DropdownItem>
                     <div role="separator" class="dropdown-divider"/>
-                    <t t-if="this.state.embeddedInfos.embeddedActions?.length">
-                        <t t-foreach="this.state.embeddedInfos.embeddedActions" t-as="action" t-key="action.id">
+                    <t t-if="this.embeddedPanelState.embeddedInfos.embeddedActions?.length">
+                        <t t-foreach="this.embeddedPanelState.embeddedInfos.embeddedActions" t-as="action" t-key="action.id">
                             <DropdownItem
                                 class="this.getDropdownClass(action)"
-                                onSelected="() => this.onEmbeddedActionClick(action)"
+                                onSelected="() => this.embeddedPanelState.onEmbeddedActionClick(action)"
                             >
                                 <span t-out="action.name" />
                             </DropdownItem>
