+++ addons/web/static/src/search/control_panel/control_panel.xml
@@ -3,22 +3,7 @@
 
     <t t-name="web.ControlPanel">
         <div class="o_control_panel d-flex flex-column gap-3 px-3 pt-2 pb-3" t-custom-ref="root" data-command-category="actions">
-            <Transition t-if="!this.env.isSmall" visible="this.state.embeddedInfos.showEmbedded" name="'o-fade'" t-slot-scope="transition" leaveDuration="500">
-                <div class="o_embedded_actions d-flex flex-wrap w-100 align-items-center justify-content-center gap-2" t-att-class="transition.className">
-                    <t t-foreach="this.state.embeddedInfos.embeddedActions" t-as="action" t-key="action.id">
-                        <t t-if="this._isEmbeddedActionVisible(action)">
-                            <button class="btn btn-secondary o_draggable"
-                                    t-att-class="{ 'active': this.state.embeddedInfos.currentEmbeddedAction?.id === action.id}"
-                                    t-on-click="() => this.onEmbeddedActionClick(action)"
-                                    t-att-data-id="action.id"
-                            >
-                                <span t-out="action.name"/>
-                            </button>
-                        </t>
-                    </t>
-                    <t t-call="web.embeddedActionsDropdown" />
-                </div>
-            </Transition>
+            <EmbeddedActionsPanel t-if="!this.env.isSmall" state="this.embeddedPanelState"/>
             <div class="o_control_panel_main d-flex flex-wrap flex-lg-nowrap justify-content-between align-items-lg-start gap-2 gap-lg-3 flex-grow-1">
                 <div class="o_control_panel_breadcrumbs d-flex align-items-center gap-1 order-0 h-lg-100">
                     <div class="o_control_panel_main_buttons d-flex gap-1 d-empty-none d-print-none text-nowrap" t-on-keydown="this.onMainButtonsKeydown">
@@ -76,15 +61,15 @@
                 </div>
 
                 <div class="o_control_panel_navigation d-flex flex-wrap flex-md-nowrap justify-content-end gap-1 gap-xl-3 order-1 order-lg-2 flex-grow-1">
-                    <t t-if="this.state.embeddedInfos.embeddedActions?.length and this.env.isSmall">
-                        <t t-call="web.embeddedActionsDropdown" />
+                    <t t-if="this.env.isSmall">
+                        <EmbeddedActionsPanel id="embedded-actions-panel-mobile" state="this.embeddedPanelState"/>
                     </t>
                     <div t-if="this.pagerProps and this.pagerProps.total > 0" class="o_cp_pager text-nowrap " role="search">
                         <Pager t-props="this.pagerProps"/>
                     </div>
-                    <button t-if="this.state.embeddedInfos.embeddedActions?.length and !this.env.isSmall" name="embedded_actions" class="btn btn-secondary"
-                            t-att-class="{active: this.state.embeddedInfos.showEmbedded}"
-                            t-on-click="this.onClickShowEmbedded"
+                    <button t-if="this.embeddedPanelState.embeddedInfos.embeddedActions?.length and !this.env.isSmall" name="embedded_actions" class="btn btn-secondary"
+                            t-att-class="{active: this.embeddedPanelState.embeddedInfos.showEmbedded}"
+                            t-on-click="() => this.embeddedPanelState.onClickShowEmbedded()"
                             title="View related documents">
                         <i class="fa fa-sliders" />
                     </button>
@@ -129,45 +114,4 @@
             </div>
         </div>
     </t>
-
-    <t t-name="web.embeddedActionsDropdown">
-        <Dropdown menuClass="'o_embedded_actions_dropdown_menu'">
-            <button name="openEmbeddedActions" class="btn btn-secondary">
-                <i class="fa fa-fw fa-sliders"/>
-            </button>
-            <t t-set-slot="content">
-                <t t-foreach="this.state.embeddedInfos.embeddedActions" t-as="action" t-key="action.id">
-                    <DropdownItem
-                        class="this.getDropdownClass(action)"
-                        onSelected="() => this.env.isSmall ? this.onEmbeddedActionClick(action) : this._setVisibility(action.id)"
-                        closingMode="'none'"
-                    >
-                        <div class="d-flex p-0 pe-1 align-items-center justify-content-between">
-                            <span t-out="action.name"/>
-                            <i t-if="action.is_deletable" class="o_icon_right btn btn-link text-danger p-0 position-absolute end-0 fa fa-trash-o" title="Delete item" t-on-click.stop="() => this.openConfirmationDialog(action)" />
-                        </div>
-                    </DropdownItem>
-                </t>
-                <div role="separator" class="dropdown-divider"/>
-                <AccordionItem class="'o_save_current_view'" description.translate="Save View">
-                    <div class="px-2 py-2">
-                        <input type="text"
-                            t-custom-ref="newActionNameRef"
-                            class="o_input mb-2"
-                            t-custom-model.trim="this.state.embeddedInfos.newActionName"
-                        />
-                        <CheckBox value="this.state.embeddedInfos.newActionIsShared" onChange.bind="this._onShareCheckboxChange">
-                            <span data-tooltip="Make this embedded action available to other users">Shared</span>
-                        </CheckBox>
-                    </div>
-                    <div class="px-3 py-2">
-                        <button class="o_save_favorite btn btn-primary w-100" t-on-click="this._saveNewAction">
-                            Save
-                        </button>
-                    </div>
-                </AccordionItem>
-            </t>
-        </Dropdown>
-    </t>
-
 </templates>
