+++ addons/web/static/src/search/control_panel/embedded_actions.xml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates xml:space="preserve">
+
+    <t t-name="web.EmbeddedActionsPanel">
+        <t t-if="!this.env.isSmall">
+            <Transition visible="this.props.state.embeddedInfos.showEmbedded" name="'o-fade'" t-slot-scope="transition" leaveDuration="500">
+                <div class="o_embedded_actions d-flex flex-wrap w-100 align-items-center justify-content-center gap-2" t-att-class="transition.className" t-custom-ref="root">
+                    <t t-foreach="this.props.state.embeddedInfos.embeddedActions" t-as="action" t-key="action.id">
+                        <t t-if="this.isEmbeddedActionVisible(action)">
+                            <button class="btn btn-secondary o_draggable"
+                                    t-att-class="{ 'active': this.props.state.embeddedInfos.currentEmbeddedAction?.id === action.id}"
+                                    t-on-click="() => this.props.state.onEmbeddedActionClick(action)"
+                                    t-att-data-id="action.id"
+                            >
+                                <span t-out="action.name"/>
+                            </button>
+                        </t>
+                    </t>
+                    <t t-call="web.EmbeddedActionsDropdown" />
+                </div>
+            </Transition>
+        </t>
+        <t t-elif="this.props.state.embeddedInfos.embeddedActions?.length">
+            <t t-call="web.EmbeddedActionsDropdown" />
+        </t>
+    </t>
+
+    <t t-name="web.EmbeddedActionsDropdown">
+        <Dropdown menuClass="'o_embedded_actions_dropdown_menu'">
+            <button name="openEmbeddedActions" class="btn btn-secondary">
+                <i class="fa fa-fw fa-sliders"/>
+            </button>
+            <t t-set-slot="content">
+                <t t-foreach="this.props.state.embeddedInfos.embeddedActions" t-as="action" t-key="action.id">
+                    <DropdownItem
+                        class="this.getDropdownClass(action)"
+                        onSelected="() => this.env.isSmall ? this.props.state.onEmbeddedActionClick(action) : this.props.state.setVisibility(action.id)"
+                        closingMode="'none'"
+                    >
+                        <div class="d-flex p-0 pe-1 align-items-center justify-content-between">
+                            <span t-out="action.name"/>
+                            <i t-if="action.is_deletable" class="o_icon_right btn btn-link text-danger p-0 position-absolute end-0 fa fa-trash-o" title="Delete item" t-on-click.stop="() => this.props.state.openConfirmationDialog(action)" />
+                        </div>
+                    </DropdownItem>
+                </t>
+                <div role="separator" class="dropdown-divider"/>
+                <AccordionItem class="'o_save_current_view'" description.translate="Save View">
+                    <div class="px-2 py-2">
+                        <input type="text"
+                            t-custom-ref="newActionNameRef"
+                            class="o_input mb-2"
+                            t-custom-model.trim="this.props.state.embeddedInfos.newActionName"
+                        />
+                        <CheckBox value="this.props.state.embeddedInfos.newActionIsShared" onChange="() => this.props.state.onShareCheckboxChange()">
+                            <span data-tooltip="Make this embedded action available to other users">Shared</span>
+                        </CheckBox>
+                    </div>
+                    <div class="px-3 py-2">
+                        <button class="o_save_favorite btn btn-primary w-100" t-on-click="(ev) => this.props.state.saveNewAction(ev, this.newActionNameRef)">
+                            Save
+                        </button>
+                    </div>
+                </AccordionItem>
+            </t>
+        </Dropdown>
+    </t>
+
+</templates>
